Arregladas las coordenadas de mario

+++ MarioBros/MarioClone/src/game/LogicalEntity.java
@@ -4,7 +4,7 @@ import factories.Sprite;
 
 public interface LogicalEntity {
     public Sprite getSprite();
-    public int getX();
-    public int getY();
+    public float getX();
+    public float getY();
     
 }
